requirements.txt
absl-py==2.1.0
accelerate==0.29.2
altair==5.3.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==3.0.0
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.3
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
diffusers==0.27.2
Distance==0.1.3
docker-pycreds==0.4.0
easydict==1.13
einops==0.7.0
executing==2.0.1
filelock==3.13.4
fire==0.7.0
fonttools==4.51.0
fsspec==2024.3.1
gdown==4.6.6
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
google-auth-oauthlib==1.0.0
grpcio==1.62.1
gudhi==3.10.1
h5py==3.11.0
huggingface-hub==0.23.1
idna==3.6
imageio==2.34.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.3
ipython==8.12.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
lazy_loader==0.4
llvmlite==0.41.1
lmdb==1.6.2
loguru==0.7.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mdurl==0.1.2
msgpack==1.0.8
munch==4.0.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.1
nltk==3.9.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.4.0.46
packaging==24.0
pandas==2.0.3
parso==0.8.4
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
prompt-toolkit==3.0.42
protobuf==4.25.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydeck==0.8.1b0
Pygments==2.17.2
pynndescent==0.5.13
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentry-sdk==1.44.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.2
streamlit==1.33.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2023.7.10
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==1.12.1
torchvision==0.13.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.41.0
typing_extensions==4.11.0
tzdata==2024.1
umap-learn==0.5.7
Unidecode==1.3.8
urllib3==2.2.1
wandb==0.14.0
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.2
zipp==3.18.1
